+++ packages/expo-modules-core/ios/Swift/AppContext.swift
@@ -28,7 +28,7 @@ public final class AppContext: NSObject {
   /**
    The legacy module registry with modules written in the old-fashioned way.
    */
-  internal weak var legacyModuleRegistry: EXModuleRegistry?
+  public weak var legacyModuleRegistry: EXModuleRegistry?
 
   internal weak var legacyModulesProxy: LegacyNativeModulesProxy?
 
@@ -77,6 +77,15 @@ public final class AppContext: NSObject {
     moduleRegistry.register(fromProvider: provider)
     return self
   }
+
+  // MARK: - UI
+
+  public func findView<ViewType>(withTag viewTag: Int, ofType type: ViewType.Type) -> ViewType? {
+    let view: UIView? = reactBridge?.uiManager.view(forReactTag: NSNumber(value: viewTag))
+    return view as? ViewType
+  }
+
+  // MARK: - Legacy modules
 
   /**
    Returns a legacy module implementing given protocol/interface.